renamed package output to artifacts
