+++ src/main/kotlin/io/thecontext/ci/artifact/DeployableArtifact.kt
@@ -1,9 +1,9 @@
-package io.thecontext.ci.output
+package io.thecontext.ci.artifact
 
 import java.io.File
 
 /**
- * A [DeployableArtifact] is a generated output (i.e. RSS Feed and show notes) that can be "deployed" with an
+ * A [DeployableArtifact] is a generated artifact (i.e. RSS Feed and show notes) that can be "deployed" with an
  * [io.thecontext.ci.deployment.DeploymentJob]
  */
 sealed class DeployableArtifact {
