+++ src/main/kotlin/io/thecontext/ci/artifact/MustacheRenderer.kt
@@ -1,4 +1,4 @@
-package io.thecontext.ci.output
+package io.thecontext.ci.artifact
 
 import com.github.mustachejava.DefaultMustacheFactory
 import java.io.BufferedReader
