+++ src/test/kotlin/io/thecontext/ci/artifact/website/WebsiteFormatterSpec.kt
@@ -1,10 +1,10 @@
-package io.thecontext.ci.output.website
+package io.thecontext.ci.artifact.website
 
 import com.greghaskins.spectrum.Spectrum
 import com.greghaskins.spectrum.dsl.specification.Specification.context
 import com.greghaskins.spectrum.dsl.specification.Specification.it
 import io.reactivex.schedulers.Schedulers
-import io.thecontext.ci.output.MustacheRenderer
+import io.thecontext.ci.artifact.MustacheRenderer
 import io.thecontext.ci.testEpisode
 import io.thecontext.ci.testPodcast
 import org.junit.runner.RunWith
